+++ lionagi/ln/__init__.py
@@ -25,15 +25,20 @@ from ._utils import (
     acreate_path,
     async_synchronized,
     coerce_created_at,
+    copy,
+    create_path,
     extract_types,
     get_bins,
     import_module,
     is_import_installed,
+    is_same_dtype,
+    is_union_type,
     load_type_from_string,
     now_utc,
     register_type_prefix,
     synchronized,
     to_uuid,
+    union_members,
 )
 from .concurrency import (
     bounded_map,
@@ -113,15 +118,20 @@ __all__ = (
     "acreate_path",
     "async_synchronized",
     "coerce_created_at",
+    "copy",
+    "create_path",
     "extract_types",
     "get_bins",
     "import_module",
     "is_import_installed",
+    "is_same_dtype",
+    "is_union_type",
     "load_type_from_string",
     "now_utc",
     "register_type_prefix",
     "synchronized",
     "to_uuid",
+    "union_members",
     "bounded_map",
     "create_task_group",
     "fail_after",
