+++ lionagi/ln/_utils.py
@@ -1,33 +1,41 @@
 import contextlib
+import copy as _copy
 import importlib
 import importlib.util
+import types
 import uuid
-from collections.abc import Awaitable, Callable
+from collections.abc import Awaitable, Callable, Mapping
 from datetime import datetime, timezone
 from functools import wraps
 from pathlib import Path as StdPath
 from types import UnionType
-from typing import Any, ParamSpec, TypeVar, Union, get_args, get_origin
+from typing import Annotated, Any, ParamSpec, TypeVar, Union, get_args, get_origin
 from uuid import UUID
 
 from anyio import Path as AsyncPath
 
 P = ParamSpec("P")
 R = TypeVar("R")
+T = TypeVar("T")
 
 __all__ = (
     "acreate_path",
     "async_synchronized",
     "coerce_created_at",
+    "copy",
+    "create_path",
     "extract_types",
     "get_bins",
     "import_module",
     "is_import_installed",
+    "is_same_dtype",
+    "is_union_type",
     "load_type_from_string",
     "now_utc",
     "register_type_prefix",
     "synchronized",
     "to_uuid",
+    "union_members",
 )
 
 
@@ -322,3 +330,112 @@ def async_synchronized(
             return await func(*args, **kwargs)
 
     return wrapper
+
+
+def copy(obj: T, /, *, deep: bool = True, num: int = 1) -> T | list[T]:
+    if num < 1:
+        raise ValueError("Number of copies must be at least 1")
+    copy_func = _copy.deepcopy if deep else _copy.copy
+    return [copy_func(obj) for _ in range(num)] if num > 1 else copy_func(obj)
+
+
+def is_same_dtype(
+    input_: list[T] | dict[Any, T],
+    dtype: type | None = None,
+    return_dtype: bool = False,
+) -> bool | tuple[bool, type | None]:
+    if not input_:
+        return (True, None) if return_dtype else True
+
+    if isinstance(input_, Mapping):
+        values_iter = iter(input_.values())
+        first_val = next(values_iter, None)
+        if dtype is None:
+            dtype = type(first_val) if first_val is not None else None
+        result = (dtype is None or isinstance(first_val, dtype)) and all(
+            isinstance(v, dtype) for v in values_iter
+        )
+    else:
+        first_val = input_[0]
+        if dtype is None:
+            dtype = type(first_val) if first_val is not None else None
+        result = all(isinstance(e, dtype) for e in input_)
+
+    return (result, dtype) if return_dtype else result
+
+
+def is_union_type(tp) -> bool:
+    """True for typing.Union[...] and PEP 604 unions (A | B)."""
+    origin = get_origin(tp)
+    return origin is Union or origin is getattr(types, "UnionType", object())
+
+
+_NoneType = type(None)
+_PEP604UnionType = getattr(types, "UnionType", None)
+
+
+def _unwrap_annotated(tp):
+    while get_origin(tp) is Annotated:
+        tp = get_args(tp)[0]
+    return tp
+
+
+def union_members(
+    tp, *, unwrap_annotated: bool = True, drop_none: bool = False
+) -> tuple[type, ...]:
+    """Return member types of a Union (typing.Union or A|B). Empty tuple if not a Union."""
+    tp = _unwrap_annotated(tp) if unwrap_annotated else tp
+    origin = get_origin(tp)
+    if origin is not Union and origin is not _PEP604UnionType:
+        return ()
+    members = get_args(tp)
+    if unwrap_annotated:
+        members = tuple(_unwrap_annotated(m) for m in members)
+    if drop_none:
+        members = tuple(m for m in members if m is not _NoneType)
+    return members
+
+
+def create_path(
+    directory: StdPath | str,
+    filename: str,
+    extension: str = None,
+    timestamp: bool = False,
+    dir_exist_ok: bool = True,
+    file_exist_ok: bool = False,
+    time_prefix: bool = False,
+    timestamp_format: str | None = None,
+    random_hash_digits: int = 0,
+) -> StdPath:
+    """Generate a file path under directory with optional timestamp and random suffix."""
+    if "/" in filename:
+        sub_dir, filename = filename.split("/")[:-1], filename.split("/")[-1]
+        directory = StdPath(directory) / "/".join(sub_dir)
+
+    if "\\" in filename:
+        raise ValueError("Filename cannot contain directory separators.")
+
+    directory = StdPath(directory)
+
+    if "." in filename:
+        name, ext = filename.rsplit(".", 1)
+    else:
+        name, ext = filename, extension
+
+    ext = f".{ext.lstrip('.')}" if ext else ""
+
+    if timestamp:
+        ts_str = datetime.now().strftime(timestamp_format or "%Y%m%d%H%M%S")
+        name = f"{ts_str}_{name}" if time_prefix else f"{name}_{ts_str}"
+
+    if random_hash_digits > 0:
+        random_suffix = uuid.uuid4().hex[:random_hash_digits]
+        name = f"{name}-{random_suffix}"
+
+    full_path = directory / f"{name}{ext}"
+
+    full_path.parent.mkdir(parents=True, exist_ok=dir_exist_ok)
+    if full_path.exists() and not file_exist_ok:
+        raise FileExistsError(f"File {full_path} already exists and file_exist_ok is False.")
+
+    return full_path
